feat: :sparkles:
Added a validator to check the data before putting it in the database

Added a validator to check the data of the body before adding/ updating a user or a rental in the database

+++ api/src/Controller/RentalController.ts
@@ -4,6 +4,10 @@ import { Advert, Prisma, PrismaClient } from "../Models/generated/prisma-client-
 
 const prisma = new PrismaClient()
 
+interface CustomRequest extends Request {
+    userId?: string;
+}
+
 /**
  * Get all rentals from the database
  * @param req: Request
@@ -13,12 +17,12 @@ async function getAllRentals(req: Request, res: Response) {
     try {
         // Get all rentals from the database and send them to the client
         const allRentals: Advert[] = await prisma.advert.findMany();
-        res.status(200).json(allRentals)
+        res.status(200).json(allRentals);
     } catch (e) {
         if (e instanceof Prisma.PrismaClientKnownRequestError) {
-            res.status(400).json({message: "Bad Request"});
+            res.status(400).json({ message: "Bad Request" });
         }
-        res.status(500).json({message: "Internal Server Error"});
+        res.status(500).json({ message: "Internal Server Error" });
     }
 }
 
@@ -28,15 +32,26 @@ async function getAllRentals(req: Request, res: Response) {
  * @param res: Response
  */
 async function getRentalsById(req: Request, res: Response) {
-    // Get the id from the request parameters
-    const id = req.params.id;
+    try {
+        // Get the id from the request parameters
+        const id = Number(req.params.id);
 
-    // Check if the rental exists
-    const rental = await prisma.advert.findUnique({ where: { id: parseInt(id) } });
-    if (rental === null) res.status(404).json({ message: "Rental Not Found" });
+        // Check if the rental exists
+        const rental = await prisma.advert.findUnique({ where: { id: id } });
+        if (rental === null) {
+            res.status(404).json({ message: "Rental Not Found" });
+            return;
+        }
 
-    // If the rental is found, send it to the client
-    res.status(200).json(rental)
+        // If the rental is found, send it to the client
+        res.status(200).json(rental);
+    } catch (e) {
+        if (e instanceof Prisma.PrismaClientKnownRequestError) {
+            res.status(400).json({ message: "Bad Request" });
+        }
+        res.status(500).json({ message: "Internal Server Error" });
+    }
+
 };
 
 /**
@@ -44,41 +59,62 @@ async function getRentalsById(req: Request, res: Response) {
  * @param req: Request
  * @param res: Response
  */
-async function createRental(req: Request, res: Response) {
-    // Get the rental's information from the request body
-    const newRental: Advert = req.body;
+async function createRental(req: CustomRequest, res: Response) {
+    try {
+        // Get the rental's informations from the request body
+        const newRental: Advert = req.body;
 
-    // Create the rental
-    const createRental: Advert = await prisma.advert.create({
-        data: newRental
-    });
+        newRental.userId = Number(req.userId);
 
-    // Send informations about the created rental to the client
-    logger.info(createRental);
-    res.status(201).json({message: "Rental created"});
+        // Create the rental
+        const createRental: Advert = await prisma.advert.create({
+            data: newRental
+        });
+
+        // Send informations about the created rental to the client
+        logger.info(createRental);
+        res.status(201).json({ message: "Rental created" });
+    } catch (e) {
+        if (e instanceof Prisma.PrismaClientKnownRequestError) {
+            res.status(400).json({ message: "Bad Request" });
+        }
+        res.status(500).json({ message: "Internal Server Error" });
+    }
+
 }
 
 /**
- * Update rental's informations by its id
+ * Update the rental's informations by its id
  * @param req: Request
  * @param res: Response
  */
 async function updateRentalById(req: Request, res: Response) {
-    // Get the id from the request parameters
-    const rentalId: number = Number(req.params.id);
+    try {
+        // Get the id from the request parameters
+        const rentalId: number = Number(req.params.id);
 
-    // Check if the rental exists
-    if (isNaN(rentalId)) res.status(400).json({message: "The rental was not found."});
+        // Check if the rental exists
+        if (await prisma.advert.findUnique({ where: { id: rentalId } }) === null) {
+            res.status(400).json({ message: "The rental was not found." });
+            return;
+        }
 
-    // If the rental exists, update it
-    const rental = await prisma.advert.update({
-        where: {id: rentalId},
-        data: req.body
-    });
+        // If the rental exists, update it
+        const rental = await prisma.advert.update({
+            where: { id: rentalId },
+            data: req.body
+        });
 
-    // Send information about the updated rental to the client
-    logger.info(rental);
-    res.status(200).json({message: "Rental Updated"});
+        // Send informations about the updated rental to the client
+        logger.info(rental);
+        res.status(200).json({ message: "Rental Updated" });
+    } catch (e) {
+        if (e instanceof Prisma.PrismaClientKnownRequestError) {
+            res.status(400).json({ message: "Bad Request" });
+        }
+        res.status(500).json({ message: "Internal Server Error" });
+    }
+
 };
 
 /**
@@ -87,20 +123,30 @@ async function updateRentalById(req: Request, res: Response) {
  * @param res: Response
  */
 async function deleteRentalById(req: Request, res: Response) {
-    // Get the id from the request parameters
-    const rentalId: number = Number(req.params.id);
+    try {
+        // Get the id from the request parameters
+        const rentalId: number = Number(req.params.id);
 
-    // Check if the rental exists
-    if (isNaN(rentalId)) res.status(400).json({message: "Bad Request"});
+        // Check if the rental exists
+        if (await prisma.advert.findUnique({ where: { id: rentalId } }) === null) {
+            res.status(400).json({ message: "Bad Request" });
+            return;
+        }
 
-    // If the rental exists, delete it
-    const deleteRental = await prisma.advert.delete({
-        where: {id: rentalId}
-    });
+        // If the rental exists, delete it
+        const deleteRental = await prisma.advert.delete({
+            where: { id: rentalId }
+        });
 
-    // Send information about the deleted rental to the client
-    logger.info(deleteRental);
-    res.status(204).json({message: "Rental Deleted"});
+        // Send informations about the deleted rental to the client
+        logger.info(deleteRental);
+        res.status(204).json({ message: "Rental Deleted" });
+    } catch (e) {
+        if (e instanceof Prisma.PrismaClientKnownRequestError) {
+            res.status(400).json({ message: "Bad Request" });
+        }
+        res.status(500).json({ message: "Internal Server Error" });
+    }
 };
 
 export { createRental, deleteRentalById, getAllRentals, getRentalsById, updateRentalById };
